application of careful casting of types

+++ app/src/main/java/dao/DAO.java
@@ -56,7 +56,7 @@ public abstract class DAO {
         assert(urlQuery != null);
         logger.log(Level.INFO,"entered in the method that makes the query");
 
-        long currentTime = Calendar.getInstance().getTime().getTime();
+        long currentTime = (long) Calendar.getInstance().getTime().getTime();
 
         Consult consult = new Consult(query,urlQuery);
         consult.exec();
